+++ tests/src/async_std.rs
@@ -78,4 +78,29 @@ mod tests {
 
         assert!(content.contains("# Get Dir"));
     }
+
+    #[async_std::test]
+    async fn test_get_dir_with_depth_limit() {
+        match GetDir::new()
+            .dir(current_dir().unwrap().join("..").join("package"))
+            .depth(1)
+            .target(Target::File(FileTarget::new("lib.rs")))
+            .run_async()
+            .await
+        {
+            | Ok(_) => panic!("Should fail"),
+            | Err(_) => (),
+        }
+
+        match GetDir::new()
+            .dir(current_dir().unwrap().join("..").join("package"))
+            .depth(2)
+            .target(Target::File(FileTarget::new("lib.rs")))
+            .run_async()
+            .await
+        {
+            | Ok(_) => (),
+            | Err(_) => panic!("Should succeed"),
+        }
+    }
 }